# distsql: improve hashjoin performance