--- a/pkg/sql/distsqlrun/hashjoiner_test.go
+++ b/pkg/sql/distsqlrun/hashjoiner_test.go
@@ -1392,10 +1392,28 @@ func BenchmarkHashJoiner(b *testing.B) {
 	st := cluster.MakeTestingClusterSettings()
 	evalCtx := tree.MakeTestingEvalContext(st)
 	defer evalCtx.Stop(ctx)
+	diskMonitor := mon.MakeMonitor(
+		"test-disk",
+		mon.DiskResource,
+		nil, /* curCount */
+		nil, /* maxHist */
+		-1,  /* increment: use default block size */
+		math.MaxInt64,
+		st,
+	)
+	diskMonitor.Start(ctx, nil /* pool */, mon.MakeStandaloneBudget(math.MaxInt64))
+	defer diskMonitor.Stop(ctx)
 	flowCtx := &FlowCtx{
-		Settings: st,
-		EvalCtx:  evalCtx,
+		Settings:    st,
+		EvalCtx:     evalCtx,
+		diskMonitor: &diskMonitor,
 	}
+	tempEngine, err := engine.NewTempEngine(base.DefaultTestTempStorageConfig(st))
+	if err != nil {
+		b.Fatal(err)
+	}
+	defer tempEngine.Close()
+	flowCtx.TempStorage = tempEngine
 
 	spec := &HashJoinerSpec{
 		LeftEqColumns:  []uint32{0},
@@ -1406,23 +1424,35 @@ func BenchmarkHashJoiner(b *testing.B) {
 	post := &PostProcessSpec{}
 
 	const numCols = 1
-	for _, numRows := range []int{0, 1 << 2, 1 << 4, 1 << 8, 1 << 12, 1 << 16} {
-		b.Run(fmt.Sprintf("rows=%d", numRows), func(b *testing.B) {
-			rows := makeIntRows(numRows, numCols)
-			leftInput := NewRepeatableRowSource(oneIntCol, rows)
-			rightInput := NewRepeatableRowSource(oneIntCol, rows)
-			b.SetBytes(int64(8 * numRows * numCols))
-			b.ResetTimer()
-			for i := 0; i < b.N; i++ {
-				// TODO(asubiotto): Get rid of uncleared state between
-				// hashJoiner Run()s to omit instantiation time from benchmarks.
-				h, err := newHashJoiner(flowCtx, spec, leftInput, rightInput, post, &RowDisposer{})
-				if err != nil {
-					b.Fatal(err)
+	for _, spill := range []bool{true, false} {
+		if spill {
+			flowCtx.testingKnobs.MemoryLimitBytes = 1
+		}
+		b.Run(fmt.Sprintf("spill=%t", spill), func(b *testing.B) {
+			for _, numRows := range []int{0, 1 << 2, 1 << 4, 1 << 8, 1 << 12, 1 << 16} {
+				if spill && numRows < 1<<8 {
+					// The benchmark takes a long time with a small number of rows and
+					// spilling, since the times change wildly. Disable for now.
+					continue
 				}
-				h.Run(context.Background(), nil /* wg */)
-				leftInput.Reset()
-				rightInput.Reset()
+				b.Run(fmt.Sprintf("rows=%d", numRows), func(b *testing.B) {
+					rows := makeIntRows(numRows, numCols)
+					leftInput := NewRepeatableRowSource(oneIntCol, rows)
+					rightInput := NewRepeatableRowSource(oneIntCol, rows)
+					b.SetBytes(int64(8 * numRows * numCols))
+					b.ResetTimer()
+					for i := 0; i < b.N; i++ {
+						// TODO(asubiotto): Get rid of uncleared state between
+						// hashJoiner Run()s to omit instantiation time from benchmarks.
+						h, err := newHashJoiner(flowCtx, spec, leftInput, rightInput, post, &RowDisposer{})
+						if err != nil {
+							b.Fatal(err)
+						}
+						h.Run(context.Background(), nil /* wg */)
+						leftInput.Reset()
+						rightInput.Reset()
+					}
+				})
 			}
 		})
 	}