- add a benchmark for the spilling hashjoiner
- reset the disk iterator rather than creating a new one every probeRow
- fix small flushing bug that caused a flush on an empty batch

```
name                                old time/op    new time/op    delta
HashJoiner/spill=true/rows=256-8      11.7ms ± 3%     2.2ms ± 3%   -81.43%  (p=0.000 n=10+8)
HashJoiner/spill=true/rows=4096-8     89.9ms ± 2%    18.4ms ± 1%   -79.58%  (p=0.000 n=10+9)
HashJoiner/spill=true/rows=65536-8     1.02s ± 2%     0.28s ± 2%   -72.16%  (p=0.000 n=9+10)

name                                old speed      new speed      delta
HashJoiner/spill=true/rows=256-8     175kB/s ± 3%   941kB/s ± 3%  +437.86%  (p=0.000 n=10+8)
HashJoiner/spill=true/rows=4096-8    364kB/s ± 2%  1780kB/s ± 2%  +389.01%  (p=0.000 n=10+10)
HashJoiner/spill=true/rows=65536-8   517kB/s ± 1%  1850kB/s ± 2%  +258.06%  (p=0.000 n=9+10)

name                                old alloc/op   new alloc/op   delta
HashJoiner/spill=true/rows=256-8       500kB ± 0%      46kB ± 0%   -90.82%  (p=0.000 n=9+9)
HashJoiner/spill=true/rows=4096-8     8.04MB ± 0%    0.79MB ± 0%   -90.16%  (p=0.000 n=9+9)
HashJoiner/spill=true/rows=65536-8     129MB ± 0%      13MB ± 0%   -90.09%  (p=0.000 n=10+9)

name                                old allocs/op  new allocs/op  delta
HashJoiner/spill=true/rows=256-8       2.88k ± 0%     0.83k ± 0%   -71.10%  (p=0.000 n=10+10)
HashJoiner/spill=true/rows=4096-8      45.4k ± 0%     12.7k ± 0%   -72.13%  (p=0.000 n=10+10)
HashJoiner/spill=true/rows=65536-8      726k ± 0%      202k ± 0%   -72.19%  (p=0.000 n=10+10)
```