@petermattis 2018-04-07
Nice speedup!

---

Review status: 0 of 4 files reviewed at latest revision, all discussions resolved, some commit checks failed.

---



*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24577#-:-L9VAtZV02M7QBMo4rzp:buerb8d)*
<!-- Sent from Reviewable.io -->