@jordanlewis 2018-04-08
PTAL.

The original patchset had a bug. You still have to recreate the underlying iterator if you've written any marks to disk. This deficiency can be solved by a cache of written marks, which we should have memory space for because once we spill to disk we abandon our memory hash table. This work is tracked in #24580.