@petermattis 2018-04-08
LGTM, though I only skimmed it. @asubiotto or @arjunravinarayan will have to give this a thorough review. 

---

Review status: 0 of 5 files reviewed at latest revision, all discussions resolved, some commit checks failed.

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 531 at r3](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24577#-L9aMKtT-Tdo1-UC_Wsx:-L9aMKtT-Tdo1-UC_Wsy:b56tcxt) ([raw file](https://github.com/cockroachdb/cockroach/blob/13002f264a7e372d24a67f57159c95d6143eae90/pkg/sql/distsqlrun/hash_row_container.go#L531)):*
> ```Go
> 	if !i.bufferedRows.Empty() {
> 		// We have to recreate our iterator if we need to flush marks to disk.
> 		// TODO(jordan): do this less by keeping a cache of written marks.
> ```

Is recreating the iterator showing up as expensive? RocksDB also has the concept of a "tailing" iterator which sees the latest database state instead of a snapshot.

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24577#-:-L9aNHN48Puym-s0qSul:b-rweoyg)*
<!-- Sent from Reviewable.io -->