@jordanlewis 2018-04-08
Review status: 0 of 5 files reviewed at latest revision, 1 unresolved discussion, some commit checks pending.

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 531 at r3](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24577#-L9aMKtT-Tdo1-UC_Wsx:-L9b5D4f9EAFSTdtUJx1:b753kei) ([raw file](https://github.com/cockroachdb/cockroach/blob/13002f264a7e372d24a67f57159c95d6143eae90/pkg/sql/distsqlrun/hash_row_container.go#L531)):*
<details><summary><i>Previously, petermattis (Peter Mattis) wrote…</i></summary><blockquote>

Is recreating the iterator showing up as expensive? RocksDB also has the concept of a "tailing" iterator which sees the latest database state instead of a snapshot.
</blockquote></details>

Yes, the reason the benchmark improved so much is that it never has to recreate/delete the iterator anymore. 

A similar benchmark of a join that needs to use marking, like EXCEPT ALL etc, would not see as big of an improvement until we complete that TODO.

I'll have to look into the tailing iterator. That might be exactly what we need.

---


*Comments from [Reviewable](https://beta.reviewable.io/reviews/cockroachdb/cockroach/24577)*
<!-- Sent from Reviewable.io -->